Move the counting.py file to agb/system
